+++ src/plugins/scmbzr/common/BzrPlugin.class.php
@@ -445,8 +445,8 @@ class BzrPlugin extends SCMPlugin {
 		}
 
 		if (! $project->enableAnonSCM()) {
-			unlink ($snapshot) ;
-			unlink ($tarball) ;
+			if (file_exists ($snapshot)) unlink ($snapshot) ;
+			if (file_exists ($tarball)) unlink ($tarball) ;
 			return false;
 		}
 
@@ -454,8 +454,8 @@ class BzrPlugin extends SCMPlugin {
 		$repo = $toprepo . '/' . $project->getUnixName() ;
 
 		if (!is_dir ($repo) || !is_file ("$repo/format")) {
-			unlink ($snapshot) ;
-			unlink ($tarball) ;
+			if (file_exists ($snapshot)) unlink ($snapshot) ;
+			if (file_exists ($tarball)) unlink ($tarball) ;
 			return false ;
 		}
