+++ src/plugins/scmdarcs/common/DarcsPlugin.class.php
@@ -381,7 +381,7 @@ class DarcsPlugin extends SCMPlugin {
 		}
 
 		if (! $project->enableAnonSCM()) {
-			unlink ($tarball) ;
+			if (file_exists ($tarball)) unlink ($tarball) ;
 			return false;
 		}
 
@@ -390,7 +390,7 @@ class DarcsPlugin extends SCMPlugin {
 		$repo = $this->getRootRepositories($project);
 
 		if (!is_dir ($repo)) {
-			unlink ($tarball) ;
+			if (file_exists ($tarball)) unlink ($tarball) ;
 			return false ;
 		}
