+++ src/plugins/scmhg/common/HgPlugin.class.php
@@ -134,7 +134,7 @@ class HgPlugin extends SCMPlugin {
 		}
 
 		if (! $project->enableAnonSCM()) {
-			unlink ($tarball) ;
+			if (file_exists ($tarball)) unlink ($tarball) ;
 			return false;
 		}
 
@@ -142,7 +142,7 @@ class HgPlugin extends SCMPlugin {
 		$repo = $toprepo . '/' . $project->getUnixName() ;
 
 		if (!is_dir ($repo)) {
-			unlink ($tarball) ;
+			if (file_exists ($tarball)) unlink ($tarball) ;
 			return false ;
 		}
